+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/utils/get-tabs-block-map.tsx
@@ -57,6 +57,7 @@ export const getTabsToBlocksMap = ({
     manageUbosBlock,
     bankAccountVerificationBlock,
     commercialCreditCheckBlock,
+    aiSummaryBlock,
   ] = blocks;
 
   const defaultTabsMap = {
@@ -65,18 +66,19 @@ export const getTabsToBlocksMap = ({
         ? caseOverviewBlock
         : []),
       ...websiteMonitoringBlock,
-      ...entityInfoBlock,
+      ...(aiSummaryBlock ? aiSummaryBlock : []),
       ...(blocksCreationParams?.workflow?.context?.pluginsOutput?.merchantScreening
         ? merchantScreeningBlock
         : []),
     ],
-    [Tab.COMPANY_INFORMATION]: [
+    [Tab.KYB]: [
+      ...kybRegistryInfoBlock,
+      ...ubosRegistryProvidedBlock,
+      ...companySanctionsBlock,
       ...entityInfoBlock,
+      ...registryInfoBlock,
       // ...mapBlock,
       ...addressWithContainerBlock,
-      ...registryInfoBlock,
-      ...kybRegistryInfoBlock,
-      ...companySanctionsBlock,
       ...bankingDetailsBlock,
       ...bankAccountVerificationBlock,
       ...commercialCreditCheckBlock,
@@ -87,9 +89,8 @@ export const getTabsToBlocksMap = ({
       ...websiteBasicRequirementBlock,
     ],
     [Tab.DOCUMENTS]: [...parentDocumentBlocks],
-    [Tab.UBOS]: [
+    [Tab.UBOS_KYC]: [
       ...ubosUserProvidedBlock,
-      ...ubosRegistryProvidedBlock,
       ...amlWithContainerBlock,
       ...manageUbosBlock,
       ...(createKycBlocks(blocksCreationParams?.workflow as TWorkflowById) || []),