+++ services/workflows-service/src/workflow-defintion/workflow-definition.repository.ts
@@ -6,6 +6,7 @@ import { GetWorkflowDefinitionListDto } from '@/workflow-defintion/dtos/get-work
 import { validateDefinitionLogic } from '@ballerine/workflow-core';
 import { Injectable } from '@nestjs/common';
 import { Prisma, PrismaClient, WorkflowDefinition } from '@prisma/client';
+import { createDemoWorkflow } from './demo-workflow/create-demo-workflow';
 
 @Injectable()
 export class WorkflowDefinitionRepository {
@@ -247,6 +248,34 @@ export class WorkflowDefinitionRepository {
         OFFSET ${dto.limit * (dto.page - 1)}
     `,
     );
+  }
+
+  async createDemoWorkflowDefinition(customerId: string) {
+    return await this.prisma.$transaction(async transaction => {
+      const customer = await transaction.customer.findUniqueOrThrow({
+        where: {
+          id: customerId,
+        },
+      });
+      const project = await transaction.project.findFirstOrThrow({
+        where: { customerId },
+        include: {
+          userToProjects: {
+            include: {
+              user: true,
+            },
+          },
+        },
+      });
+
+      const demoEnv = {
+        customer,
+        project,
+        user: project.userToProjects[0]?.user,
+      };
+
+      await createDemoWorkflow(customer, demoEnv, transaction);
+    });
   }
 
   async findByWorkflowRuntimeDataId(workflowRuntimeDataId: string, projectIds: TProjectIds) {