+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # web-ui-sdk
 
+## 1.5.79
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.78
+
+## 1.5.78
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.77
+
+## 1.5.77
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.76
+
 ## 1.5.76
 
 ### Patch Changes